+++ crates/wdl-analysis/src/analyzer.rs
@@ -53,6 +53,7 @@ use crate::queue::AnalyzeRequest;
 use crate::queue::CallHierarchyRequest;
 use crate::queue::CodeLensRequest;
 use crate::queue::CompletionRequest;
+use crate::queue::DeleteRequest;
 use crate::queue::DocumentSymbolRequest;
 use crate::queue::FindAllReferencesRequest;
 use crate::queue::FoldingRangeRequest;
@@ -64,11 +65,12 @@ use crate::queue::InlayHintsRequest;
 use crate::queue::NotifyChangeRequest;
 use crate::queue::NotifyIncrementalChangeRequest;
 use crate::queue::OutgoingCallsRequest;
-use crate::queue::RemoveRequest;
 use crate::queue::RenameRequest;
 use crate::queue::Request;
 use crate::queue::SemanticTokenRequest;
 use crate::queue::SignatureHelpRequest;
+use crate::queue::SwapValidatorRequest;
+use crate::queue::UnrootDocumentsRequest;
 use crate::queue::WorkspaceSymbolRequest;
 use crate::rayon::RayonHandle;
 
@@ -514,8 +516,13 @@ where
         let inner_config = config.clone();
         let inner_resolution = resolution.clone();
         let handle = std::thread::spawn(move || {
-            let queue =
-                AnalysisQueue::new(inner_config, tokio, inner_resolution, progress, validator);
+            let queue = AnalysisQueue::new(
+                inner_config,
+                tokio,
+                inner_resolution,
+                progress,
+                Arc::new(validator),
+            );
             queue.run(rx);
         });
 
@@ -525,6 +532,30 @@ where
             config,
             resolution,
         }
+    }
+
+    /// Replace the current validator function.
+    ///
+    /// This will mark all documents for re-analysis.
+    pub async fn swap_validator<Validator>(&self, validator: Validator) -> Result<()>
+    where
+        Validator: Fn() -> crate::Validator + Send + Sync + 'static,
+    {
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::SwapValidator(SwapValidatorRequest {
+                validator: Arc::new(validator),
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!("failed to send request to analysis queue because the channel has closed")
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!("failed to receive response from analysis queue because the channel has closed")
+        })?;
+
+        Ok(())
     }
 
     /// Adds a document to the analyzer. Document can be a local file or a URL.
@@ -669,11 +700,36 @@ where
     /// the analyzer, those documents will be removed.
     ///
     /// Documents are only removed when not referenced from importing documents.
-    pub async fn remove_documents(&self, documents: Vec<Url>) -> Result<()> {
-        // Send the remove request to the queue
+    /// To forcefully delete the documents from the graph, use
+    /// [`Self::delete_documents()`].
+    pub async fn unroot_documents(&self, documents: Vec<Url>) -> Result<()> {
+        // Send the unroot request to the queue
         let (tx, rx) = oneshot::channel();
         self.sender
-            .send(Request::Remove(RemoveRequest {
+            .send(Request::UnrootDocuments(UnrootDocumentsRequest {
+                documents,
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!("failed to send request to analysis queue because the channel has closed")
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!("failed to receive response from analysis queue because the channel has closed")
+        })?;
+
+        Ok(())
+    }
+
+    /// Deletes the specified documents from the analyzer.
+    ///
+    /// This differs from [`Self::unroot_documents()`], as a deletion will occur
+    /// even if the document(s) are referenced in other documents.
+    pub async fn delete_documents(&self, documents: Vec<Url>) -> Result<()> {
+        // Send the delete request to the queue
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::Delete(DeleteRequest {
                 documents,
                 completed: tx,
             }))
@@ -1508,7 +1564,7 @@ workflow test {
 
         // Remove the documents by directory
         analyzer
-            .remove_documents(vec![
+            .unroot_documents(vec![
                 path_to_uri(dir.path()).expect("should convert to URI"),
             ])
             .await
@@ -1842,5 +1898,64 @@ workflow run {
             resolver.max_active.load(Ordering::SeqCst) > 1,
             "symbolic imports should materialize concurrently"
         );
+    }
+
+    #[tokio::test]
+    async fn it_deletes_documents() {
+        let dir = TempDir::new().expect("failed to create temporary directory");
+        let foo = dir.path().join("foo.wdl");
+        fs::write(
+            &foo,
+            r#"version 1.1
+import "bar.wdl"
+
+workflow test {
+    call bar.test
+}
+"#,
+        )
+        .expect("failed to create test file");
+
+        let bar = dir.path().join("bar.wdl");
+        fs::write(
+            &bar,
+            r#"version 1.1
+workflow test {}
+"#,
+        )
+        .expect("failed to create test file");
+
+        // Add both documents to the analyzer
+        let analyzer = Analyzer::default();
+        analyzer
+            .add_directory(dir.path())
+            .await
+            .expect("should add documents");
+
+        // Analyze the documents
+        let results = analyzer.analyze(()).await.unwrap();
+        assert_eq!(results.len(), 2);
+        assert!(results[0].document.diagnostics().next().is_none());
+        assert!(results[1].document.diagnostics().next().is_none());
+
+        // Now delete bar.wdl, which foo.wdl depends on.
+        //
+        // Unlike removal, this should *force* the deletion of bar.wdl in the graph (and
+        // thus cause errors in foo.wdl)
+        fs::remove_file(&bar).expect("should delete file");
+        analyzer
+            .delete_documents(vec![path_to_uri(&bar).expect("should convert to URI")])
+            .await
+            .unwrap();
+
+        // Now foo.wdl should error
+        let results = analyzer.analyze(()).await.unwrap();
+        assert_eq!(results.len(), 1);
+
+        let has_import_failed_diagnostic = results[0]
+            .document
+            .diagnostics()
+            .any(|d| d.message().contains("failed to import `bar.wdl`"));
+        assert!(has_import_failed_diagnostic);
     }
 }